% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_emissions.R
\name{report_emissions}
\alias{report_emissions}
\title{Report Emissions.}
\usage{
report_emissions(emispath = "post", path_df = "post/df",
  path_streets = "post/streets", path_grids = "post/grids")
}
\arguments{
\item{emispath}{Character; path to post .rds}

\item{path_df}{Character; path to data-frames .rds}

\item{path_streets}{Character; path to streets .rds}

\item{path_grids}{Character; path to grids .rds}
}
\value{
a list
}
\description{
\code{\link{report_traffic}} runs traffic functions.
}
\examples{
\dontrun{
name <- file.path(tempdir())
vein::inventory(name)
source(paste0(name, "/main.R"))
loe <- report_emissions()
names(loe)
}
}
