% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_speed.R
\name{report_speed}
\alias{report_speed}
\title{Report Speed.}
\usage{
report_speed(speed, net)
}
\arguments{
\item{speed}{path to speed.rds}

\item{net}{path to net.rds}
}
\value{
a list
}
\description{
\code{\link{report_traffic}} runs traffic functions.
}
\examples{
\dontrun{
vein::inventory(file.path(tempdir()), rush.hour = TRUE)
source(paste0(file.path(tempdir(), "main.R")))
net <- readRDS("network/net.rds")
net <- net[, c("ldv", "hdv", "ps")]
speed <- readRDS("network/speed.rds")
los <- report_speed(speed, net)
names(los)
}
}
