% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_traffic.R
\name{report_traffic}
\alias{report_traffic}
\title{Report Traffic.}
\usage{
report_traffic(net, width = 3000, ..count..)
}
\arguments{
\item{net}{Spatial Feature net with numeric columns}

\item{width}{Numeric; Widht if cell grid.}

\item{..count..}{ignored.}
}
\value{
a list
}
\description{
\code{\link{report_traffic}} runs traffic functions.
}
\examples{
\dontrun{
vein::inventory(file.path(tempdir()),rush.hour = T)
source(paste0(file.path(tempdir(), "main.R")))
net <- readRDS("network/net.rds")
net <- net[, c("ldv", "hdv", "ps")]
lot <- report_traffic(net)
names(lot)
}
}
