% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_black.R
\name{theme_black}
\alias{theme_black}
\title{Theme for ggplot2.}
\usage{
theme_black(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\value{
theme black
}
\description{
theme black ...
}
\note{
based on https://gist.github.com/jslefche/eff85ef06b4705e6efbc#file-theme_black-r
}
\examples{
\dontrun{
# do not run
}
}
