% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vein_report.R
\name{vein_report}
\alias{vein_report}
\title{Inventory function.}
\usage{
vein_report(author = "me", path = getwd(), traffic = TRUE, speed = TRUE,
  ef = "scaled", emissions = TRUE)
}
\arguments{
\item{author}{Character; name of the author of the report.}

\item{path}{Character; path where to create report.Rmd.}

\item{traffic}{Logical; if TRUE writes traffic functions.}

\item{speed}{Logical; if TRUE writes speed functions.}

\item{ef}{Character; 'local', 'speed' or 'scaled'.}

\item{emissions}{Logical; if TRUE writes emissions functions.}
}
\value{
report.Rmd.
}
\description{
\code{\link{vein_report}} produces an structure of directories and scripts
in order to run vein. It is required to know the vehicular composition of the
fleet.
}
\examples{
{
name <- file.path(tempdir())
vein::inventory(name)
source(paste0(name, "/main.R"))
vein_report(author = "me", path = name)
rmarkdown::render(paste0(name, "/report.Rmd"))
}
}
