% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_speed.R
\name{report_speed}
\alias{report_speed}
\title{Report Speed.}
\usage{
report_speed(speed = "network/speed.rds", net = "network/net.rds")
}
\arguments{
\item{speed}{path to speed.rds}

\item{net}{path to net.rds}
}
\value{
a list
}
\description{
\code{\link{report_traffic}} runs traffic functions.
}
\examples{
\dontrun{
vein::inventory(file.path(tempdir()))
source(paste0(file.path(tempdir(), "main.R")))
los <- report_speed()
names(los)
}
}
