% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_traffic.R
\name{report_traffic}
\alias{report_traffic}
\title{Report Traffic.}
\usage{
report_traffic(net = "network/net.rds", width = 1/102, ..count..)
}
\arguments{
\item{net}{path to net.rds}

\item{width}{Numeric; Widht if cell grid.}

\item{..count..}{ignored.}
}
\value{
a list
}
\description{
\code{\link{report_traffic}} runs traffic functions.
}
\examples{
\dontrun{
vein::inventory(file.path(tempdir()))
source(paste0(file.path(tempdir(), "main.R")))
lot <- report_traffic()
names(lot)
}
}
